refactor(tui): extract shared text utilities into textutil.go
- Move softWrapLines, truncate, padRight, padEmoji to tui/textutil.go
- Replace all direct runewidth.Truncate calls with shared truncate()
- Remove unused runewidth imports from list.go and template_editor.go

Closes #335
